# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Franchise Development OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginWithPassword
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Signed in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '400':
          description: Missing email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Account deactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Clear the current session
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /auth/accept-invite:
    post:
      tags: [Auth]
      operationId: acceptInvitation
      summary: Claim an invite link by setting name and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptance'
      responses:
        '200':
          description: Invite accepted; user is now signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '400':
          description: Invalid invite, name, or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: changeMyPassword
      summary: Change your own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
      responses:
        '200':
          description: Password changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '400':
          description: Current password incorrect or new password invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /impersonate/start:
    post:
      tags: [Auth]
      operationId: startImpersonation
      summary: Begin acting as another user (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpersonationTarget'
      responses:
        '200':
          description: Now acting as the target user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '400':
          description: Invalid target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not permitted to impersonate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /reports/definitions:
    get:
      tags: [Reporting]
      operationId: listReportDefinitions
      summary: List saved report definitions (presets + custom)
      responses:
        "200":
          description: List of report definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportDefinition"
    post:
      tags: [Reporting]
      operationId: createReportDefinition
      summary: Create a saved report definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDefinitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDefinition"
  /reports/run:
    post:
      tags: [Reporting]
      operationId: runReport
      summary: Run a report over canonical CRM data (internal aggregation only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportSpec"
      responses:
        "200":
          description: Report result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    InvitationAcceptance:
      type: object
      required: [token, name, password]
      properties:
        token:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
...
    ImpersonationTarget:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          minLength: 1
    UserRole:
      type: string
      enum: [member, lead, admin, vp, super_admin, observer]
    ActingIdentity:
      type: object
      description: The real super admin behind an impersonated session.
      required: [id, name, avatarColor, role, profileImageUrl]
      properties:
        id:
          type: string
        name:
          type: string
        avatarColor:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        profileImageUrl:
          type: [string, 'null']
    AppUser:
      type: object
      required: [id, name, email, role, avatarColor]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: [string, 'null']
        role:
          $ref: "#/components/schemas/UserRole"
        avatarColor:
          type: string
        profileImageUrl:
          type: [string, 'null']
        hasLogin:
          type: boolean
          description: True when the member has set a password and can sign in.
        isActive:
          type: boolean
          description: False when the user has been deactivated and can no longer log in.
        permissions:
          type: array
          description: Resolved permission keys for this user's role (populated on /me).
          items:
            $ref: "#/components/schemas/PermissionKey"
        impersonating:
          type: boolean
          description: True when a super admin is currently acting as this user.
        realUser:
          oneOf:
            - $ref: "#/components/schemas/ActingIdentity"
            - type: 'null'
          description: The real super admin driving an impersonated session, else null.
...
    PermissionKey:
      type: string
      enum:
        - impersonate_users
        - manage_users
        - manage_permissions
        - view_audit_log
        - manage_access_requests
        - manage_invitations
        - manage_groups
        - assign_tasks
        - approve_completion
        - edit_initiatives
        - delete_initiatives
        - manage_targets
        - view_reports
        - run_reports
        - manage_metrics
        - manage_integrations
...
    ReportSource:
      type: string
      enum: [opportunities, campaigns, accounts, people, email_activity]
    ReportFilter:
      type: object
      required: [field, op, value]
      properties:
        field:
          type: string
        op:
          type: string
          enum: [eq, neq, in, notIn, gt, gte, lt, lte]
        value: {}
    ReportMetric:
      type: object
      required: [agg]
      properties:
        agg:
          type: string
          enum: [count, sum, avg]
        field:
          type: [string, 'null']
          description: Numeric field to aggregate. Ignored for count.
        label:
          type: [string, 'null']
    ReportChartType:
      type: string
      enum: [bar, line, pie, area, table]
    ReportSpec:
      type: object
      required: [sources, dimensions, metrics]
      properties:
        sources:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ReportSource"
        dimensions:
          type: array
          description: Fields to group rows by; each becomes a dimension column.
          items:
            type: string
        metrics:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ReportMetric"
        filters:
          type: array
          items:
            $ref: "#/components/schemas/ReportFilter"
        groupBy:
          type: [string, 'null']
          description: Optional primary grouping field (defaults to the first dimension).
        chartType:
          oneOf:
            - $ref: "#/components/schemas/ReportChartType"
            - type: 'null'
    ReportColumn:
      type: object
      required: [key, label, kind]
      properties:
        key:
          type: string
        label:
          type: string
        kind:
          type: string
          enum: [dimension, metric]
    ReportResult:
      type: object
      required: [spec, generatedAt, columns, rows]
      properties:
        spec:
          $ref: "#/components/schemas/ReportSpec"
        generatedAt:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ReportColumn"
        rows:
          type: array
          description: Each row is keyed by column key; cells are string, number, or null.
          items:
            type: object
            additionalProperties:
              oneOf:
                - type: string
                - type: number
                - type: 'null'
    ReportDefinition:
      type: object
      required: [id, name, spec, isPreset, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        spec:
          $ref: "#/components/schemas/ReportSpec"
        isPreset:
          type: boolean
        createdById:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ReportDefinitionInput:
      type: object
      required: [name, spec]
      properties:
        name:
          type: string
          minLength: 1
        spec:
          $ref: "#/components/schemas/ReportSpec"